[BUGFIX] Fix invalid format of delta (metric) value
By default Golang will use a decimal exponent when converting larger float value to a string. This issue was not caught due to small values being used in testing.

This bug is fixed by explicitly converting the float64 value to a string and then specifying that we will not use exponential formatting. There is also a test added to catch this issue.