move tests to their own package (godspeed_test)
Doing these tests within the main package was a bit of a rookie mistake. I'm looking to convert these tests to use the go-check package as it offers a better interface for testing. So, before doing this big change, I thought it made sense to move the tests to the `godspeed_test` package. In addition, I moved the helper methods out in to their own package as well.