Update imports, .travis.yml, and copyright headers for new repo
This change updates the repository to ensure that all of the import paths point
to the new repository. This is needed as we fork `PagerDuty/godspeed` and look
to become the canonical repository.

In addition to changing the import paths, this also includes a copyright for Tim
Heckman on files I've modified as part of this changes.

Finally, this updates the `.travis.yml` file to test against modern versions of
the Go runtime.

Signed-off-by: Tim Heckman <t@heckman.io>